+++ src/Location.h
@@ -16,7 +16,7 @@ public:
      * Creates a new Location, if no arguments are specified we create
      * a 0,0,0 Location, wich is the invalid Location.
      */
-    Location(Uint32 x = 0, Uint32 z = 0, Uint32 y = 0);
+    Location(Uint16 x = 0, Uint16 z = 0, Uint16 y = 0);
 
     /**
      * Virtual destructor.
@@ -30,8 +30,8 @@ public:
      * @since pre-alpha
      * @return The X coordinate of the Location.
      */
-    Uint32 getX() { SDL_LockMutex(mxCoords);
-                    Uint32 X = mX;
+    Uint16 getX() { SDL_LockMutex(mxCoords);
+                    Uint16 X = mX;
                     SDL_UnlockMutex(mxCoords);
                     return X;}
 
@@ -42,7 +42,7 @@ public:
      * @since pre-alpha
      * @param newX The new X coordinate.
      */
-    void setX(Uint32 newX) {SDL_LockMutex(mxCoords);
+    void setX(Uint16 newX) {SDL_LockMutex(mxCoords);
                             mX = newX;
                             SDL_UnlockMutex(mxCoords);}
 
@@ -53,8 +53,8 @@ public:
      * @since pre-alpha
      * @return The Y coordinate of the Location.
      */
-    Uint32 getY() { SDL_LockMutex(mxCoords);
-                    Uint32 Y = mY;
+    Uint16 getY() { SDL_LockMutex(mxCoords);
+                    Uint16 Y = mY;
                     SDL_UnlockMutex(mxCoords);
                     return Y;}
 
@@ -65,7 +65,7 @@ public:
      * @since pre-alpha
      * @param newY The new Y coordinate.
      */
-    void setY(Uint32 newY) {SDL_LockMutex(mxCoords);
+    void setY(Uint16 newY) {SDL_LockMutex(mxCoords);
                             mY = newY;
                             SDL_UnlockMutex(mxCoords);}
 
@@ -76,8 +76,8 @@ public:
      * @since pre-alpha
      * @return The Z coordinate of the Location.
      */
-    Uint32 getZ() { SDL_LockMutex(mxCoords);
-                    Uint32 Z = mZ;
+    Uint16 getZ() { SDL_LockMutex(mxCoords);
+                    Uint16 Z = mZ;
                     SDL_UnlockMutex(mxCoords);
                     return Z;}
 
@@ -88,7 +88,7 @@ public:
      * @since pre-alpha
      * @param newZ The new Z coordinate.
      */
-    void setZ(Uint32 newZ) {SDL_LockMutex(mxCoords);
+    void setZ(Uint16 newZ) {SDL_LockMutex(mxCoords);
                             mZ = newZ;
                             SDL_UnlockMutex(mxCoords);}
 
@@ -101,7 +101,7 @@ public:
      * @param newY New Y coordinate.
      * @param newZ New Z coordinate.
      */
-    void setXYZ(Uint32 newX, Uint32 newY, Uint32 newZ) {SDL_LockMutex(mxCoords);
+    void setXYZ(Uint16 newX, Uint16 newY, Uint16 newZ) {SDL_LockMutex(mxCoords);
                                                         mX = newX;
                                                         mY = newY;
                                                         mZ = newZ;
@@ -116,7 +116,7 @@ public:
      * @param y Returned Y coordinate.
      * @param z Returned Z coordinate.
      */
-    void getXYZ(Uint32 &x, Uint32 &y, Uint32 &z) {  SDL_LockMutex(mxCoords);
+    void getXYZ(Uint16 &x, Uint16 &y, Uint16 &z) {  SDL_LockMutex(mxCoords);
                                                     x = mX;
                                                     y = mY;
                                                     z = mZ;
@@ -140,9 +140,9 @@ public:
 
 private:
 
-    Uint32 mX;          /**< My X coordinate. */
-    Uint32 mY;          /**< My Y coordinate. */
-    Uint32 mZ;          /**< My Z coordinate. */
+    Uint16 mX;          /**< My X coordinate. */
+    Uint16 mY;          /**< My Y coordinate. */
+    Uint16 mZ;          /**< My Z coordinate. */
     SDL_mutex* mxCoords;/**< Mutex to prevents concurrency during reading/writing of the coords. */
 };
